+++ src/config/index.ts
@@ -17,3 +17,5 @@
  *   }
  */
 export const fonts = ['inter', 'manrope', 'system'] as const
+
+export const LANG = 'EN'